+++ uCNC/src/interface/grbl_interface.h
@@ -132,6 +132,7 @@ extern "C"
 #define GRBL_PINS_STATES (GRBL_SYSTEM_CMD + 13)
 #define GRBL_SEND_SYSTEM_INFO (GRBL_SYSTEM_CMD + 14)
 #define GRBL_SEND_SYSTEM_INFO_EXTENDED (GRBL_SYSTEM_CMD + 15)
+#define GRBL_PRINT_PARAM (GRBL_SYSTEM_CMD + 16)
 
 #define GRBL_SYSTEM_CMD_EXTENDED (GRBL_SYSTEM_CMD + 20)
 #define GRBL_SYSTEM_CMD_EXTENDED_UNSUPPORTED 253
